# feat: Configure GitHub Pages deployment with PWA and enhanced caching