--- a/app/root.tsx
+++ b/app/root.tsx
@@ -10,6 +10,7 @@ import {
 import type { Route } from './+types/root';
 import './app.css';
 import { ThemeProvider } from './contexts/ThemeContext';
+import { ErrorBoundary as AppErrorBoundary } from './utils/errorBoundary';
 
 export const links: Route.LinksFunction = () => [
   { rel: 'preconnect', href: 'https://fonts.googleapis.com' },
@@ -34,7 +35,9 @@ export function Layout({ children }: { children: React.ReactNode }) {
         <Links />
       </head>
       <body>
-        <ThemeProvider>{children}</ThemeProvider>
+        <AppErrorBoundary>
+          <ThemeProvider>{children}</ThemeProvider>
+        </AppErrorBoundary>
         <ScrollRestoration />
         <Scripts />
       </body>