--- a/app/utils/analytics.ts
+++ b/app/utils/analytics.ts
@@ -70,4 +70,4 @@ declare global {
     dataLayer: any[];
     gtag: (...args: any[]) => void;
   }
-} 
+}