--- a/public/.nojekyll
+++ b/public/.nojekyll
@@ -1,0 +1,2 @@
+# This file tells GitHub Pages not to process this site with Jekyll
+# This is needed for React Router to work properly