… .github/workflows/deploy.yml
… README.md
… app/utils/errorBoundary.tsx
… app/utils/production.ts
… app/utils/routing.ts
… app/utils/seo.ts
… env.example
… package.json
… public/404.html
… public/index.html
… public/manifest.json
… public/robots.txt
… public/sitemap.xml
… public/sw.js
… react-router.config.ts
… vite.config.ts
… yarn.lock